# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AMHARC Match Capture — Local Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches/{matchId}/events:
    get:
      operationId: getMatchEvents
      tags: [events]
      summary: Get match events
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchEvent"
    post:
      operationId: createMatchEvent
      tags: [events]
      summary: Create a match event
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchEventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchEvent"
  /matches/{matchId}/events/{eventId}:
    put:
      operationId: updateMatchEvent
      tags: [events]
      summary: Update a match event
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchEventUpdate"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchEvent"
    delete:
      operationId: deleteMatchEvent
      tags: [events]
      summary: Delete a match event
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /matches/{matchId}/events/undo:
    post:
      operationId: undoLastEvent
      tags: [events]
      summary: Undo the last event
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Undo result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
...
  /matches/{matchId}/export:
    post:
      operationId: exportMatch
      tags: [exports]
      summary: Export match data package
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"
...
components:
  schemas:
...
    MatchEvent:
      type: object
      required: [eventId, matchId, eventType, period, matchClockSeconds, recordingElapsedSeconds, systemTimestamp, source, reviewStatus, createdAt, updatedAt]
      properties:
        eventId:
          type: string
        matchId:
          type: string
        eventType:
          type: string
          enum: [match-start, match-end, period-start, period-end, half-time-start, half-time-end, score, goal, point, two-point-score, shot, kick-out, puck-out, turnover, free, mark, sideline-cut, hook, block, substitution, card, injury, major-incident, highlight, technical-issue, custom-event]
        team:
          type: ["string", "null"]
          enum: [home, away, null]
        playerId:
          type: ["string", "null"]
        playerNumber:
          type: ["integer", "null"]
        period:
          type: integer
        matchClockSeconds:
          type: integer
        recordingElapsedSeconds:
          type: integer
        systemTimestamp:
          type: string
          format: date-time
        source:
          type: string
          enum: [operator-ui, stream-deck, joystick, system, imported, api, automatic]
        operator:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        scoreBefore:
          type: ["string", "null"]
        scoreAfter:
          type: ["string", "null"]
        clipRequested:
          type: boolean
        reviewStatus:
          type: string
          enum: [unreviewed, reviewed, corrected, rejected, flagged]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MatchEventInput:
      type: object
      required: [eventType, period, matchClockSeconds, recordingElapsedSeconds, source]
      properties:
        eventType:
          type: string
          enum: [match-start, match-end, period-start, period-end, half-time-start, half-time-end, score, goal, point, two-point-score, shot, kick-out, puck-out, turnover, free, mark, sideline-cut, hook, block, substitution, card, injury, major-incident, highlight, technical-issue, custom-event]
        team:
          type: string
          enum: [home, away]
        playerId:
          type: string
        playerNumber:
          type: integer
        period:
          type: integer
        matchClockSeconds:
          type: integer
        recordingElapsedSeconds:
          type: integer
        source:
          type: string
          enum: [operator-ui, stream-deck, joystick, system, imported, api, automatic]
        note:
          type: string
        clipRequested:
          type: boolean
    MatchEventUpdate:
      type: object
      properties:
        team:
          type: string
          enum: [home, away]
        playerId:
          type: string
        playerNumber:
          type: integer
        note:
          type: string
        reviewStatus:
          type: string
          enum: [unreviewed, reviewed, corrected, rejected, flagged]
        clipRequested:
          type: boolean
        matchClockSeconds:
          type: integer
...
    ExportRequest:
      type: object
      required: [formats]
      properties:
        formats:
          type: array
          items:
            type: string
            enum: [json, csv, manifest, technical-log]
        includeEventTypes:
          type: array
          items:
            type: string
    ExportResult:
      type: object
      required: [success, exportDirectory, files]
      properties:
        success:
          type: boolean
        exportDirectory:
          type: string
        files:
          type: array
          items:
            type: string
        manifestVersion:
          type: integer
        message:
          type: ["string", "null"]
    OperationResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
        error:
          type: ["string", "null"]